+++ src/basic/components/product/ProductList.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 
 import { ProductCard } from "./ProductCard";
-import { Product } from "../../types";
+import { Product } from "../../../types";
 
 interface ProductListProps {
   products: Product[];
